more

+++ frontend/src/utils/usePostComment.js
@@ -1,20 +1,15 @@
-import { useContext, useState } from "react"
-import actions from "../api"
-import { SongDataContext } from "../contexts/SongData"
-import TheContext from "../contexts/TheContext"
+import { useContext, useState } from 'react'
+import actions from '../api'
+import { SongDataContext } from '../contexts/SongData'
+import TheContext from '../contexts/TheContext'
 
 export default function usePostComment() {
   const { user } = useContext(TheContext)
-  const { allSongComments, setAllSongComments } = useContext(SongDataContext)
+  const { homeFeedSongs, setHomeFeedSongs } = useContext(SongDataContext)
 
-  const initialComments = {
-    USERS_COMMENT: false,
-    USER_COMMENT_TO_DELETE: null,
-    TOTAL_COMMENTS: null
-  }
-  const [comments, setComments] = useState(initialComments)
+  const addComment = (songId, commentString) => {
+    if (songId == null || commentString == null) return
 
-  const handlePostComment = (songId, commentString) => {
     actions
       .addComment({
         songId: songId,
@@ -22,62 +17,60 @@ export default function usePostComment() {
         date: new Date(),
       })
       .then(res => {
-        console.log(`ADDED a COMMENT: ---`, res.data.comment, `--- to ${res.data.song.name}'s song_comments: `, res.data.song.song_comments)
+        console.log(
+          `ADDED a COMMENT: ---`,
+          res.data.comment,
+          `--- to ${res.data.song.name}'s song_comments: `,
+          res.data.song.song_comments,
+        )
         const songComments = res.data.song.song_comments
         const commentToDelete = res.data.comment
 
-        setComments(prevComments => ({
-            ...prevComments,
-            TOTAL_COMMENTS: songComments,
-            USER_COMMENT_TO_DELETE: commentToDelete,
-        }))
-        
-        let newCommentsArr = allSongComments.map(comment => {
-            if (comment.songId === songId) {
-              return {...comment, comments: songComments }
-            } else {
-              return comment
-            }
-          })
+        // let updateFeed = homeFeedSongs.map(each => {
+        //   if (each.song._id === songId) {
+        //     return { ...each, song: { ...each.song, song_comments: songComments } }
+        //   } else return each
+        // })
+        // setHomeFeedSongs(updateFeed)
 
-        setAllSongComments(newCommentsArr)
+        setHomeFeedSongs(prev =>
+          prev.map(each => {
+            if (each.song._id === songId)
+              return { ...each, song: { ...each.song, song_comments: songComments } }
+            else return each
+          }),
+        )
       })
       .catch(console.error)
   }
 
-  const handleDeleteComment = (songId, toDelete) => {
-    if (user?._id !== toDelete.user._id) return
+  const deleteComment = (songId, toDelete) => {
+    if (user?._id !== toDelete.user._id || songId == null || toDelete == null) return
+
     actions
       .deleteComment({ songId: songId, commentToDelete: toDelete })
       .then(res => {
-        console.log(`DELETED a COMMENT: ---`, res.data.comment, `--- from ${res.data.song.name}'s song_comments: `, res.data.song.song_comments)
+        console.log(
+          `DELETED a COMMENT: ---`,
+          res.data.comment,
+          `--- from ${res.data.song.name}'s song_comments: `,
+          res.data.song.song_comments,
+        )
         const songComments = res.data.song.song_comments
 
-        setComments(prevComments => ({
-          ...prevComments,
-          TOTAL_COMMENTS: songComments,
-          USER_COMMENT_TO_DELETE: null
-        }))
-
-        let newCommentsArr = allSongComments.map(comment => {
-          if (comment.songId === songId) {
-            return {...comment, comments: songComments }
-          } else {
-            return comment
-          }
+        let updateFeed = homeFeedSongs.map(each => {
+          if (each.song._id === songId) {
+            return { ...each, song: { ...each.song, song_comments: songComments } }
+          } else return each
         })
 
-        setAllSongComments(newCommentsArr)
-      
+        setHomeFeedSongs(updateFeed)
       })
       .catch(console.error)
   }
 
   return {
-    handlePostComment,
-    handleDeleteComment,
-    initialComments,
-    comments,
-    setComments
+    addComment,
+    deleteComment,
   }
 }
